Fixed moveAction not detecting getting blocked
moveAction wasn't detecting properly when the unit was blocked.

To fix this, when the move has started (this means that the order move
has been sent to the unit) on each call of onUnit the following
condition is checked:

if (unit.getOrder().equals(Order.PlayerGuard) || unit.isAttacking() ||
unit.isStartingAttack())

to ensure that the unit is currently moving or not.